+++ junosterraform/go_template_2.py
